Remove obsolete fixmelpa references
The fixmelpa submodule and all references were already removed in
9a8eb69e27, but the references in sometimesfood-packages.el were
erroneously reintroduced in f83d672630.

This commit drops support for Emacs 24 (which is currently broken
anyway, due to the removal of the fixmelpa submodule). Supporting
Emacs 24 is too much of a hassle, since I do not use it myself
anymore.